We don't need no MONKEY-TYPING
Originally, the idea was to augment Str/List with appropriate pack/unpack
methods.  But as jnthn++ pointed out, no two precomped modules augmenting
the same class, can be loaded at the same time.  So we're shelving the
idea of augmenting classes for now, and leave that to the user if they
really feel inclined to do so.